fix(tenant-registration): route invites to tenant domain
